Removes the old type inference backend
RIP...

To eulogize: we don't actually need this anymore; it's encoding
didn't scale, and a more direct translation better suited to automated
verification is now available in FlowInference. So rather than
have to maintain the one backend we actually want, and one that no one uses
just so the program compiles, let aGiLeLy and aggressively delete code
no one uses.