Adds omit havoc optimization
This commit removes access paths that correspond to havoced locations.
Under relaxed mode, these locations can never be read, and never provide
any useful information. It is therefore sound to remove them.

This removal is not complete, in particular there are some places in mu
relations where computing the set of havoc paths is quite difficult. In that case,
the current approach is sound, havoc paths that have not been omitted
are still havoced as usual.